requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
bcrypt
python-jose
fastapi
pydantic
passlib
python-multipart
python-dotenv
smtplib
email-validator